Separate PUT & POST methods and check update diff
PUT is used for new kitties and is idempotent (except that the generated name
will be different).
POST is used for updating existing kitties and is not idempotent, because the
latest update is applied as a diff on top of whatever is already in the
kitty.

This means that if client A puts 100 into the account, then client B adds 50,
and then client A adds 20, the total in the account will be 170. The server
and client A will have the correct value; client B is currently not updated
automatically until it pushes an update to the server or reloads.